Float param support
...and some saved patterns.

+++ server/sockets/control.py
@@ -34,7 +34,7 @@ def sanitizeParams(params):
     if type(value) == dict and 'RGBValues' in value:
       params[key] = Color((value['RGBValues']))
     elif is_number(value):
-      params[key] = int(value)
+      params[key] = float(value)
   return params
 
 class DiscoControlProtocol(WebSocketServerProtocol):
